agentic-malloy: extract table-spec.ts (break the layer-build↔glossary cycle)
Move TableSpec/TableInput/NormTable + quoteDuckRef/safeTableName/normalizeTables
into src/table-spec.ts so glossary.ts can use them without importing layer-build
(which will soon import glossary). layer-build re-exports them for existing
importers. No behavior change; tsc clean, 106/106.
